date in events: remove time if any
Some versions of SDWorx calendar have a time but we don't want it, it's
in the middle of other events.

Also, if you ask to take part of the day or part of half a day, the time
might not be correct, e.g. if you take 6h out of 8, are you going to
work in the morning or the afternoon?

So we think it is better to have "whole day events" in all cases, easier
to distinct from meetings.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>