long events: summary: set days + AM/PM
When SDWorx is generating a calendar with time, they also display '8h' for one
day and even if it covers more than 1 day. Easier to read '1D' than '8h' or
the proper amount of days instead of '8h'.

Also, when removing the (possibly) wrong time, it is still interesting to
extract AM/PM info.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>