hourly: add '?' if more than a day
There is a bug in SDWorx where events initially on a few hours are
merged with others and cover more than one day.

For example if someone takes 3h on the afternoon on two consecutive
days, a single event will be created with the start time of the first
one and the end of the second one, e.g. from today at 13:00 to tomorrow
at 16:00. That's wrong but we don't have more details to fix this event.

Just to say that the info we got are probably not OK, we add a '?' in
the title and increment the number of days.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>